requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.7
fastapi==0.103.2
filelock==3.12.4
greenlet==3.0.0
h11==0.14.0
identify==2.5.30
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
platformdirs==3.11.0
pre-commit==3.4.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
ruff==0.0.292
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
virtualenv==20.24.5

requests~=2.31.0